+++ frontend/src/pages/dashboard/UsersView.vue
@@ -1,0 +1,197 @@
+<template>
+	<div class="flex items-center space-x-3">
+		<FormControl v-model="search" :placeholder="__('Search')" class="w-80">
+			<template #prefix>
+				<FeatherIcon name="search" class="w-4 text-gray-600" />
+			</template>
+		</FormControl>
+		<FormControl
+			v-model="role"
+			:placeholder="__('Member Role')"
+			class="w-40"
+			type="select"
+			:options="ROLE_OPTIONS"
+			@update:model-value="members.reload"
+		/>
+	</div>
+	<ListView
+		v-if="members?.data"
+		class="flex-1"
+		:columns="[{ label: __('User'), key: 'user' }]"
+		:rows="members.data"
+		:options="LIST_OPTIONS"
+		row-key="name"
+	>
+		<ListHeader />
+		<ListRows>
+			<template v-if="members.data.length">
+				<ListRow
+					v-for="row in members.data"
+					:key="row.name"
+					:row="row"
+					:class="{
+						'cursor-pointer rounded hover:bg-gray-50': row.name !== tenantOwner.data,
+					}"
+					@click="openAccount(row.name)"
+				>
+					<div class="grid grid-cols-3">
+						<div class="flex items-center space-x-2">
+							<Avatar :image="row.user_image" :label="row.full_name" size="lg" />
+							<div class="text-sm">
+								<p class="font-medium">{{ row.full_name }}</p>
+								<p class="mt-0.5 text-gray-600">
+									{{ row.name }}
+								</p>
+							</div>
+						</div>
+						<div class="mx-auto flex items-center">
+							<Badge
+								v-if="row.is_admin"
+								:theme="row.name === tenantOwner.data ? 'orange' : 'blue'"
+								:label="__(row.name === tenantOwner.data ? 'Owner' : 'Admin')"
+							/>
+						</div>
+						<div class="ml-auto flex items-center">
+							<Dropdown
+								v-if="row.name !== tenantOwner.data"
+								:options="dropdownOptions(row.name, row.is_admin)"
+								:button="{
+									icon: 'more-horizontal',
+									variant: 'ghost',
+								}"
+								@click.stop
+							/>
+						</div>
+					</div>
+				</ListRow>
+			</template>
+			<ListEmptyState v-else />
+		</ListRows>
+	</ListView>
+	<Dialog v-model="showRemoveMember" :options="removeMemberOptions" />
+	<MailAccountModal v-model="showMailAccount" :account-i-d="selectedMailAccount" />
+</template>
+
+<script setup lang="ts">
+import { computed, inject, ref } from 'vue'
+import { watchDebounced } from '@vueuse/core'
+import {
+	Avatar,
+	Badge,
+	Dialog,
+	Dropdown,
+	FeatherIcon,
+	FormControl,
+	ListEmptyState,
+	ListHeader,
+	ListRow,
+	ListRows,
+	ListView,
+	createResource,
+} from 'frappe-ui'
+
+import { raiseToast } from '@/utils'
+import MailAccountModal from '@/components/Modals/MailAccountModal.vue'
+
+const user = inject('$user')
+
+const search = ref('')
+const role = ref<'Mail User' | 'Mail Admin' | ''>('')
+const showRemoveMember = ref(false)
+const memberToBeRemoved = ref('')
+const showMailAccount = ref(false)
+const selectedMailAccount = ref('')
+
+const tenantOwner = createResource({
+	url: 'frappe.client.get_value',
+	makeParams: () => ({
+		doctype: 'Mail Tenant',
+		fieldname: 'user',
+		filters: user.data?.tenant,
+		as_dict: false,
+	}),
+	auto: true,
+	cache: ['mailTenantOwner', user.data?.tenant],
+})
+
+const members = createResource({
+	url: 'mail.api.admin.get_tenant_members',
+	makeParams: () => ({ tenant: user.data?.tenant, search: search.value, role: role.value }),
+	auto: true,
+	cache: ['mailTenantMembers', user.data?.tenant, search.value, role.value],
+})
+
+watchDebounced(() => search.value, members.reload, { debounce: 500 })
+
+const reloadMembers = () => members.reload()
+defineExpose({ reloadMembers })
+
+const editAdminRole = createResource({
+	url: 'frappe.client.set_value',
+	makeParams: ({ name, is_admin }: { name: string; is_admin: 0 | 1 }) => ({
+		doctype: 'Mail Tenant Member',
+		name: name,
+		fieldname: 'is_admin',
+		value: is_admin,
+	}),
+	onSuccess: () => {
+		raiseToast(__('Role updated successfully'))
+		members.reload()
+	},
+	onError: (error) => raiseToast(error.messages[0], 'error'),
+})
+
+const removeMember = createResource({
+	url: 'frappe.client.delete',
+	makeParams: () => ({ doctype: 'Mail Tenant Member', name: memberToBeRemoved.value }),
+	onSuccess: () => {
+		raiseToast(__('Member removed successfully'))
+		showRemoveMember.value = false
+		members.reload()
+	},
+	onError: (error) => raiseToast(error.messages[0], 'error'),
+})
+
+const openAccount = (account: string) => {
+	if (account !== tenantOwner.data) {
+		selectedMailAccount.value = account
+		showMailAccount.value = true
+	}
+}
+
+const dropdownOptions = (name: string, isAdmin: 0 | 1) => [
+	{
+		label: isAdmin ? __('Remove Admin') : __('Make Admin'),
+		icon: isAdmin ? 'shield-off' : 'shield',
+		onClick: () => editAdminRole.submit({ name, is_admin: !isAdmin }),
+	},
+	{
+		label: __('Remove Member'),
+		icon: 'user-x',
+		onClick: () => {
+			memberToBeRemoved.value = name
+			showRemoveMember.value = true
+		},
+	},
+]
+
+const removeMemberOptions = computed(() => ({
+	title: __('Remove Member'),
+	message: __(`Are you sure you want to remove member ${memberToBeRemoved.value}?`),
+	icon: { name: 'alert-triangle', appearance: 'warning' },
+	actions: [{ label: __('Confirm'), variant: 'solid', onClick: removeMember.submit }],
+}))
+
+const LIST_OPTIONS = {
+	selectable: false,
+	showTooltip: false,
+	rowHeight: 50,
+	emptyState: { description: __('No members found.') },
+}
+
+const ROLE_OPTIONS = [
+	{ label: '', value: '' },
+	{ label: __('Mail User'), value: 'Mail User' },
+	{ label: __('Mail Admin'), value: 'Mail Admin' },
+]
+</script>
